Desktop/khedma/11/smartup-backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
Pillow==8.0.1
python-decouple==3.3
gunicorn==20.0.4
drf-spectacular==0.16.0
django-absoluteuri==1.3.0
dj-database-url==0.5.0
django-guardian==2.3.0
whitenoise==5.2.0
psycopg2-binary==2.8.6
pyfcm==1.4.7
django-extensions==3.0.9
django-cleanup==5.1.0
bcrypt==3.2.0
bugsnag==3.6.0
factory-boy==3.1.0
Faker==4.17.1
django-cors-headers==3.5.0
model-bakery==1.2.1
celery==4.4.7
django-celery-beat==2.1.0
django-celery-results==2.0.0
flower==0.9.5
parameterized==0.7.4
unittest-xml-reporting==3.0.4
freezegun==1.0.0
django-macros==0.4.0
PyJWT==1.7.1
qrcode==7.3
pdfkit==0.6.1
django-url-filter==0.3.15
django-import-export==2.6.1
rules==3.0
httmock==1.4.0
djangorestframework-api-key==2.2.0
coverage==6.4.4
drf-pdf==0.2.0
docxtpl==0.16.4
haversine==2.7.0
django-redis==5.2.0
requests==2.25.1 
urllib3==1.26.5